requirements.txt
fastapi
PyJWT
passlib
starlette
pydantic[email]
python-jose
cryptography
fastapi-mcp
psycopg[binary]
SQLAlchemy
sqlmodel
requests
numpy
scikit-learn
Pillow
shapely
pyproj
networkx
osmnx
soundfile
librosa
pytest-asyncio
pytest-timeout
google-genai
bs4
pypdf
pandas
nltk
